back-fastapi/requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
databases==0.7.0
fastapi==0.97.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
pydantic==1.10.9
python-dotenv==1.0.0
requests==2.31.0
setuptools==65.5.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0